+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/OrchestratorMappings.kt
@@ -43,6 +43,13 @@ class OrchestratorMappings(
 ) {
     private val logger = KotlinLogging.logger { }
 
+    fun toCreateRequest(entity: BusinessPartnerDb): TaskCreateRequestEntry{
+        return TaskCreateRequestEntry(
+            recordId = entity.sharingState.orchestratorRecordId?.toString(),
+            businessPartner = toOrchestratorDto(entity)
+        )
+    }
+
     fun toOrchestratorDto(entity: BusinessPartnerDb): BusinessPartner {
         val postalAddress = toPostalAddress(entity)